+++ src/pages/User/User.tsx
@@ -1,21 +1,15 @@
-import { useEffect } from 'react';
-import toast from 'react-hot-toast';
+import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 
 import Button from '@/components/atoms/Button/Button';
+import UserBanner from '@/components/Banner/UserBanner/UserBanner';
 import MyTripCard from '@/components/Cards/MyTripCard';
 import UserInfo from '@/components/UserInfo/UserInfo';
-import { useAuthStore } from '@/store/AuthStore';
+import { useGetApi } from '@/hooks/useGetAPI';
 import type { Trip } from '@/types/api/Trip.type';
+import type { User } from '@/types/api/User.type';
 
 import styles from './User.module.scss';
-
-interface UserPageProps {
-  tab: 'upcoming' | 'past';
-  upcoming: Trip[];
-  past: Trip[];
-  userInfo: { name: string; phone: string; email: string } | null;
-}
 
 const tabInfo = {
   upcoming: {
@@ -28,22 +22,41 @@ const tabInfo = {
   }
 } as const;
 
-const UserPage = ({ tab, upcoming, past, userInfo }: UserPageProps) => {
+type TripTab = 'upcoming' | 'past';
+
+const UserPage = () => {
   const navigate = useNavigate();
-  const { isLoggedIn } = useAuthStore();
+  const [tab, setTab] = useState<TripTab>('upcoming');
 
-  // 인증 체크
-  useEffect(() => {
-    if (!isLoggedIn) {
-      toast.error('로그인이 필요한 페이지입니다.');
-      navigate('/login', { replace: true });
-    }
-  }, [isLoggedIn, navigate]);
+  // 데이터 패칭
+  const {
+    data: tripsResponse,
+    isLoading: tripsLoading,
+    isError: tripsError
+  } = useGetApi('/users/me/trips');
+  const {
+    data: userInfoResponse,
+    isLoading: userInfoLoading,
+    isError: userInfoError
+  } = useGetApi('/users/me');
 
-  // 로그인되지 않은 경우 렌더링하지 않음
-  if (!isLoggedIn) {
-    return null;
-  }
+  // 데이터 가공
+  const tripsData: Trip[] = tripsResponse?.data ?? [];
+  const userInfoData: User | null = userInfoResponse?.data ?? null;
+  const today = new Date();
+  const upcoming = Array.isArray(tripsData)
+    ? tripsData.filter((trip) => new Date(trip.end_date) >= today)
+    : [];
+  const past = Array.isArray(tripsData)
+    ? tripsData.filter((trip) => new Date(trip.end_date) < today)
+    : [];
+  const userInfo = userInfoData
+    ? {
+        name: userInfoData.name,
+        phone: userInfoData.phoneNumber,
+        email: userInfoData.email
+      }
+    : null;
 
   const { title, emptyMsg } = tabInfo[tab];
   const trips = tab === 'upcoming' ? upcoming : past;
@@ -52,35 +65,51 @@ const UserPage = ({ tab, upcoming, past, userInfo }: UserPageProps) => {
     navigate(`/product/${productId}`);
   };
 
+  if (tripsLoading || userInfoLoading) {
+    return <div className={styles.loading}>로딩 중...</div>;
+  }
+  if (tripsError || userInfoError) {
+    return <div className={styles.error}>데이터를 불러오는 중 오류가 발생했습니다.</div>;
+  }
+
   return (
     <>
-      <div className={styles.sectionTitle}>{title}</div>
-      <div key={tab} className={styles.userPageWrapper}>
-        {trips.length === 0 ? (
-          <div className={styles.emptyMsg}>{emptyMsg}</div>
-        ) : (
-          trips.map((trip) => (
-            <div
-              key={`${trip.productId}_${trip.start_date}_${trip.end_date}`}
-              className={styles.userTripWrapper}
-            >
-              <MyTripCard trip={trip} />
-              {userInfo && (
-                <UserInfo name={userInfo.name} phone={userInfo.phone} email={userInfo.email} />
-              )}
-              <div className={styles.tripButtonWrapper}>
-                <Button
-                  variant="sm"
-                  color="gray"
-                  style={{ margin: '16px 0 0 0' }}
-                  onClick={() => handleProductPageNavigation(trip.productId)}
-                >
-                  해당 상품 페이지로
-                </Button>
+      <UserBanner
+        tab={tab}
+        setTab={setTab}
+        upcomingCount={upcoming.length}
+        pastCount={past.length}
+        userName={userInfo?.name || ''}
+      />
+      <div className={styles.userContentWrapper}>
+        <div className={styles.sectionTitle}>{title}</div>
+        <div key={tab} className={styles.userPageWrapper}>
+          {trips.length === 0 ? (
+            <div className={styles.emptyMsg}>{emptyMsg}</div>
+          ) : (
+            trips.map((trip) => (
+              <div
+                key={`${trip.productId}_${trip.start_date}_${trip.end_date}`}
+                className={styles.userTripWrapper}
+              >
+                <MyTripCard trip={trip} />
+                {userInfo && (
+                  <UserInfo name={userInfo.name} phone={userInfo.phone} email={userInfo.email} />
+                )}
+                <div className={styles.tripButtonWrapper}>
+                  <Button
+                    variant="sm"
+                    color="gray"
+                    style={{ margin: '16px 0 0 0' }}
+                    onClick={() => handleProductPageNavigation(trip.productId)}
+                  >
+                    해당 상품 페이지로
+                  </Button>
+                </div>
               </div>
-            </div>
-          ))
-        )}
+            ))
+          )}
+        </div>
       </div>
     </>
   );
